+++ src/types/report.dto.ts
@@ -1,12 +1,3 @@
-export interface Report {
-  imageId: number;
-  lastReportAt: string;
-  reportCount: number;
-  tags: { tagId: number; tagName: string }[];
-}
-
-export interface ReportDetail {
-  reportDate: string;
-  reportUserEmail: string;
-  tags?: { tagId: number; tagName: string }[];
-}
+import { Report, ReportDetail } from "./report";
+export type GetReportResponse = Report[];
+export type GetReportDetailResponse = ReportDetail[];
